refactor: view 관련 클래스를 하나의 패키지로 이동
