+++ mvc/src/main/java/com/interface21/webmvc/servlet/view/ModelAndView.java
@@ -1,5 +1,7 @@
-package com.interface21.webmvc.servlet;
+package com.interface21.webmvc.servlet.view;
 
+import jakarta.servlet.http.HttpServletRequest;
+import jakarta.servlet.http.HttpServletResponse;
 import java.util.Collections;
 import java.util.HashMap;
 import java.util.Map;
@@ -29,5 +31,9 @@ public class ModelAndView {
 
     public View getView() {
         return view;
+    }
+
+    public void render(final HttpServletRequest request, final HttpServletResponse response) throws Exception {
+        view.render(model, request, response);
     }
 }
